+++ HISTORY.rst
@@ -15,6 +15,7 @@ History
     * Refactor all mocks used for resource tests (rudaporto).
     * Added new test to check RESTService.collection_get method (rudaporto).
     * Move database fixtures to conftest and use Base sqlalchemy from briefy.common (rudaporto).
+    * Fix: validate column filters before trying to access attr to create the filter (rudaporto).
 
 
 2.0.6 (2017-07-27)